docs: Added Philips AirPurifier AC3858/50 model (#43)
The Philips AirPurifier AC3858/50 is supported by this card. The only culprits are, that there is a problem (sometimes) with setting the mode from mode dropdown - you need to set it twice to work. Also, the selected mode doesn't display right in the toolbar. I'll try to sort this out somehow.